Move to a src dir and separate test dir
